dropdown temporary changes and other design fixes.

+++ 360Accounting.Web/Controllers/CodeCombinitionController.cs
@@ -26,7 +26,7 @@ namespace _360Accounting.Web.Controllers
         {
             CodeCombinitionCreateViewModel model =  new CodeCombinitionCreateViewModel();
             model.SegmentList = AccountHelper.GetSegmentListForCodeCombination(sobId);
-            model.SOBId = Convert.ToInt32(sobId);
+            model.SOBId = Convert.ToInt64(sobId);
             return PartialView("_Create", model);
         }
